SUPPORT2-160 Service-ek létrehozása
Added Services for the project and modified some data models.
Added necessary gradle imports for endpoints and modified the main to work like an api server.

+++ src/main/java/org/example/model/Submission.java
@@ -30,6 +30,6 @@ public class Submission {
     private Boolean acceptance;
 
     @Column(value = "SUBMITTERID")
-    private Integer submitterid;
+    private Long submitterid;
 
 }
